Extract a context module
